Update procedures.js
change number type from float to double for make block function

+++ blocks_vertical/procedures.js
@@ -709,7 +709,7 @@ Blockly.ScratchBlocks.ProcedureUtils.addStringExternal = function() {
   Blockly.WidgetDiv.hide(true);
   this.procCode_ =  this.procCode_.replace(/"/g, "");
   this.procCode_ = this.procCode_ + ' %n';
-  this.displayNames_.push('float');
+  this.displayNames_.push('double');
   this.argumentIds_.push(Blockly.utils.genUid());
   this.argumentDefaults_.push(0);
   this.updateDisplay_();
